move exception interfaces to their own namespace
