+++ packages/EasyErrorHandler/src/Interfaces/Exceptions/LogLevelAwareExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyErrorHandler\Interfaces\Exceptions;
+
+interface LogLevelAwareExceptionInterface
+{
+    public function getLogLevel(): int;
+}
