+++ packages/EasyErrorHandler/src/Interfaces/Exceptions/StatusCodeAwareExceptionInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyErrorHandler\Interfaces;
+namespace EonX\EasyErrorHandler\Interfaces\Exceptions;
 
 interface StatusCodeAwareExceptionInterface
 {
