+++ packages/EasyErrorHandler/src/Interfaces/Exceptions/TranslatableExceptionInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyErrorHandler\Interfaces;
+namespace EonX\EasyErrorHandler\Interfaces\Exceptions;
 
 interface TranslatableExceptionInterface
 {
