--- packages/EasyErrorHandler/src/Interfaces/LogLevelAwareExceptionInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EonX\EasyErrorHandler\Interfaces;
-
-interface LogLevelAwareExceptionInterface
-{
-    /**
-     * Returns the log level of an exception.
-     */
-    public function getLogLevel(): string;
-}
